normalize structure
