+++ components/model-endpoints-filter/index.ts
@@ -1,7 +1,8 @@
-export { ModelEndpointsFilterWithData } from './ModelEndpointsFilterWithData'
-export { ModelFilterControls } from './ModelFilterControls'
-export { ModelFilterResults } from './ModelFilterResults'
-export { FilterCheckbox } from './FilterCheckbox'
-export { filterModels } from './utils'
+export { ModelEndpointsFilter } from './model-endpoints-filter'
+export { ModelFilterControls } from './model-filter-controls'
+export { ModelFilterResults } from './model-filter-results'
+export { ModelFilterItem } from './results/model-filter-item'
+export { filterModels, hasActiveFilters, urlStateToFilterState } from './filter'
 
 export type { FilterState, SortOption, SortDirection, FilterResult } from './types'
+export { SORT_CONFIG, SORT_OPTIONS } from './types'
